refactor(library/HTTP): dissolve "HTTPResponseError.ts"
